requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifultable==0.8.0
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
Django==4.1.1
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
entrypoints==0.4
executing==1.1.1
ghasedakpack==0.1.8
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
isodate==0.6.1
jdatetime==4.1.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
lxml==4.9.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
tzlocal==2.1
urllib3==1.26.12
wcwidth==0.2.5
websocket-client==1.4.1
zeep==4.2.0
